+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+* Fixed bug where framerate was not locked at max 60 Hz
+
 ## [0.2.1] - 2020-02-13
 ### Added
 * Gamepads connected after the game started can now be used. The same goes for
@@ -30,7 +34,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.1.1] - 2019-12-21
 ### Fixed
-* Corrected the refresh rate to work for a 60hz monitor
+* Corrected the refresh rate to work for a 60 Hz monitor
 
 ## [0.1.0] - 2019-12-20
 ### Added